add ping reponse and fix parse error
