+++ tests/samples/ping_success.json
@@ -1,0 +1,12 @@
+{
+  "name" : "Scorcher",
+  "cluster_name" : "elasticsearch",
+  "version" : {
+    "number" : "5.0.0-alpha3",
+    "build_hash" : "cad959b",
+    "build_date" : "2016-05-26T08:25:57.564Z",
+    "build_snapshot" : false,
+    "lucene_version" : "6.0.0"
+  },
+  "tagline" : "You Know, for Search"
+}
